requirements.txt
supabase
python-dotenv
fastapi
uvicorn
python-jose[cryptography]
pydantic
pydantic-settings
groq
python-dateutil
crewai
crewai-tools
langchain-community
langchain-groq
litellm